Refactor utils directory
