+++ package.json
@@ -35,7 +35,6 @@
   "dependencies": {
     "JSONStream": "^1.1.1",
     "babel-cli": "^6.14.0",
-    "babel-core": "^6.4.5",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.4.3",
@@ -43,37 +42,28 @@
     "batch-stream": "^0.1.3",
     "bluebird": "^3.3.5",
     "body-parser": "^1.15.1",
+    "cache-manager": "^2.2.0",
+    "connect-timeout": "^1.8.0",
     "cors": "^2.8.1",
     "csv-stream": "^0.1.3",
     "ejs": "^2.4.1",
-    "eslint": "^2.9.0",
-    "event-stream": "^3.3.4",
     "express": "^4.13.4",
-    "flat": "^2.0.1",
     "foreman": "^2.0.0",
-    "html": "0.0.10",
     "hull": "0.10.2",
-    "jwt-simple": "^0.5.0",
     "kue": "^0.11.1",
     "kue-ui": "^0.1.0",
     "librato-node": "^4.1.0",
-    "lint": "^1.1.2",
     "lodash": "^4.12.0",
     "moment": "^2.14.1",
     "newrelic": "^1.28.3",
-    "object-mapper": "^3.0.1",
-    "omit-deep-lodash": "^0.8.0",
     "passport-intercom": "0.0.2",
     "promise-streams": "^1.0.1",
     "raven": "^0.12.1",
     "request": "^2.72.0",
-    "request-promise": "^3.0.0",
-    "simple-oauth2": "^0.7.0",
     "superagent": "^2.2.0",
     "superagent-prefix": "0.0.2",
     "superagent-promise-plugin": "^3.2.0",
     "supply": "0.0.4",
-    "tar-stream": "^1.5.2",
     "urijs": "^1.18.0"
   },
   "devDependencies": {
@@ -83,7 +73,6 @@
     "eslint-config-airbnb-base": "^3.0.1",
     "eslint-plugin-import": "^1.8.0",
     "mocha": "^2.5.2",
-    "proxyquire": "^1.7.10",
     "sinon": "^1.17.4"
   }
 }
