+++ server/controller/notif-handlers.js
@@ -1,6 +1,7 @@
 import Promise from "bluebird";
 import _ from "lodash";
-import BatchSyncHandler from "../util/batch-sync-handler";
+
+import BatchSyncHandler from "../util/handler/batch-sync";
 
 export default class NotifHandlers {
