+++ server/util/app/web.js
@@ -3,15 +3,12 @@ import path from "path";
 import { renderFile } from "ejs";
 import timeout from "connect-timeout";
 
-import TokenMiddleware from "../util/middleware/token";
-
 export default function WebApp() {
   const app = express();
 
   // the main responsibility of following timeout middleware
   // is to make the web app respond always in time
   app.use(timeout("25s"));
-  app.use(TokenMiddleware);
   app.engine("html", renderFile);
 
   app.set("views", path.resolve(__dirname, "..", "..", "views"));
