+++ server/util/handler/exit.js
