+++ server/util/hull-agent/index.js
@@ -1,4 +1,5 @@
 import _ from "lodash";
+
 import ExtractAgent from "./extract-agent";
 
 export default class HullAgent {
@@ -47,35 +48,5 @@ export default class HullAgent {
 
   getSegments() {
     return this.hullClient.get("/segments");
-  }
-
-  updateUser(member) {
-    const traits = this.getUserTraitsForMember(member);
-    this.hullClient.logger.info("updateUser", traits.email);
-    const { email, unique_email_id } = traits;
-    const ident = { email };
-    if (unique_email_id) {
-      ident.anonymous_id = `mailchimp:${unique_email_id}`;
-    }
-
-    return this.hullClient
-      .as(ident)
-      .traits(traits, { source: "mailchimp" });
-  }
-
-  getExtractFields() {
-    const traits = this.mailchimpFields.map(path => {
-      const trait = _.last(path.split("."));
-      return `traits_mailchimp/${trait}`;
-    });
-    const props = [
-      "traits_mailchimp/import_error",
-      "traits_mailchimp/last_activity_at",
-      "id",
-      "email",
-      "first_name",
-      "last_name"
-    ];
-    return props.concat(traits);
   }
 }
