+++ server/util/middleware/token.js
@@ -4,7 +4,7 @@
  * @param  {Object}   res
  * @param  {Function} next
  */
-export default function handle(req, res, next) {
+export default function tokenMiddleware(req, res, next) {
   if (req.query && req.query.hullToken) {
     req.hull = req.hull || {};
     req.hull.token = req.query.hullToken;
