+++ server/util/router/static.js
@@ -1,7 +1,7 @@
 import express from "express";
 import path from "path";
 
-export default function ({ Hull }) {
+export default function StaticRouter({ Hull }) {
   const router = express.Router();
   const { Routes } = Hull;
   const { Readme, Manifest } = Routes;
